freection-2 Cannot see the details of a task
Adding a Task page.
It is accessible from Whats New, To Do, and Follow Up.
Currently accessible by clicking the Thing name, which is a link.

A few things to notice:

1) There's a fair amount of code duplication here.
Both in logic (routing logic for instance) and in GUI.

2) Each Thing now has a payload and a type.
Right now, all Things are of type TASK.
We'd change that when we go deeper into the New logic.

3) Event has an eventType instead of just type.
That could be helpful when an event would have also a thingType property, in the future.

4) Better to reset the DB.

6) I had to use the full blown lodash for the 'remove' function.
I think we should be able to bundle our app properly, even when using lodash...