freection-12 I see comments on the same task separately in Whats New
A small improvement to the way the message is previewed in the To Do and Follow Up pages.
The components should dynamically decided if there are new Comments that should be displayed, so that while writing Comments in the UI, the To Do and Follow Up views are updated accordingly.

This is still not perfect, as we have to wait for the polling to occur in order to update the components.
It's not worth fixing IMO, since it's too much effort which is useless when we move to pubsub.