Bug fix - assigning token when user is not authenticated.
In the static server, when the page is served at the beginning, we also add a token for the push service.
If the user had not been authenticated to begin with, adding this extra token is irrelevant.
I was trying to sign a JWT for the user, accessing the user object in the request, while there was no user object since the user had not been authenticated at all.