Bug fix in immutable.
Trying to setOrPush to an array that does not exist, resulted in error.
This can happen when the showTask reducer tries to update a comment in a Thing, but no Thing is currently open (therefore the state is an empty object, and the array of comments cannot be found in it).