Bug fix - wrong display of comments aggregation in Whats New.
It resulted in a wrong usage of lodash merge - it mutates the original object.
I use clone in order to avoid that.

Notice that I also changed the way we require lodash in this file.
It mixes with the bug fix change, but I figured I can still do them together...