Changing thing.comments to thing.events.
This is kinda big - now all events are returned on the Thing.
The UI should filter them out as it sees fit.

It will enable us to add PING as an event by itself, and the UI will decide how to display it (currently, most probably as a regular comment).

In the future, we'd also be able to show an activity log of the Thing.