Adding the Unread Emails page.
This is a WIP - the bodies of the unread emails are not yet shown, and there is no Email Page yet.

There's also no periodic update of the unread emails list.

Getting the unread emails is done by a separate server - mail-fetch.
It has HTTP API, and opens an IMAP connection for every user.