Fixing connection pool bug.
Since creating an IMAP connection is an async action, when several requests are made for the same user, which should be using the very same IMAP connection - they might attempt to create several connections (the map of connections is not filled with the newly created connection before new requests arrive).
We need to use a smarter pool, that would have new requests wait until the connection is ready.
Each request should also release the connection.