freection-47 thing page doesn't get invalidated on error
When there's an error, the page gets invalidated, and the data is re-fetched.
It happens for a get action, and for an async command action (i.e. Done, Dismiss, etc.)
This results in a bit of an ugly flick of the screen, or worse - a long wait for the new fetch to arrive.

Not sure this is an ideal way.