freection-57 sometimes by mistake I send two things
We now disable the Send button if the async action against the server does not return in a few millis.

I also change the action buttons of Thing Page do behave the same (almost).
The status of the Thing will be changed only after the async action returns from the server.
In the meantime, the buttons would be disabled.
Problem is that they turn disabled right away, with no delay (like the Send button).