freection-69 I see the Message Box in the Github Integration page
In order to really support this, I made the refactor to the Message Box.
There are now two boxes - NewThingBox and CommentThingBox.
There could be also NewMailBox and ReplyMailBox.

It also means that right now, no box at all is displayed for the Email Page (which makes sense, since it would not have worked).

The style of the boxes is still shared - we can split it too.