freection-95 I can't enter a message when hitting Dismiss
This is kind of an attempt to address the issue.
The styling is not exactly like in the mockups - there are some open questions there.

Also, when dismissing from the To Do list - the last comment is not displayed for the Inline Message.
The reason is that we don't have easy access to the last comment - we only have the Thing in hand, not a specific notification/event/comment.