Styling fix - Message Box was responsive in a weird way.
Would sometimes be too wide (when resizing down the screen) and sometimes too narrow (when resizing up the screen).

I now make it stretch according to screen size, while the only important thing is to keep its padding.

Also, the message box in non-full-mode gets extra padding, so that it would have a similar width to when it's displayed in full-mode.