freection-164 I cannot add email to my todo list
This is an initial version of the Do Email from Gmail flow.

Supported:
1) Installing the extension locally.
2) The Do button appears in Gmail.
3) It accesses Freection and creates a Thing corresponding the email.
4) The cookie seems to be sent on the REST call to Freection - but there are many caveats.

Not supported:
1) Installing the extension from the store.
2) Better UI and UX for the button in Gmail.
3) Implementing the Do flow in Freection better - creating a full-blown Thing for the email might create complications for now, so we need to simplify it.
4) For instance, while you can reply on the email-thing from Freection - you won't get additional messages on the thread in What's New Things. You will get them as regular emails (if emails tab is open), and not  on the Thing itself...
5) Regarding cookie and token - if the token is saved as a cookie on the browser, it will be sent when the extension calls Freection.
But there are two problematic cases: when the token is not in the cookie, for whatever reason; and when the user has more than one Freection account.
These might be lesser problems for now, and we might need to come with an API for login which returns a token, in order to better support these cases.