freection-164 I cannot add email to my todo list
Make sure the extension works with freection.com. and not localhost.

For dev mode, one would have to find a solution - perhaps change hosts file in order to direct freection.com to localhost, or simply change the code...
Once we have a proper build for the extension, we can also think about building a prod and dev extensions, which will both be installed in our Chrome.