freection-164 I cannot add email to my todo list
I change the way the chrome extension code is loaded.
It is now loaded from a JS file we serve.
In order to support this, we must support HTTPS.

In order to support HTTPS in dev mode, we now work against a URL provided by ngrok, which redirects to our localhosts.
We have 2 ngrok endpoints now, one for each localhost...

HTTPS should also be supported in production for this to work.