freection-164 I cannot add email to my todo list
Introducing an API for creating a Thing from an Email, w/o the need for IMAP access to the email.
The needed email data is provided to the API call.
The Email Thing that's created is in fact a link to the email in Gmail.

Currently the API is open for any authenticated user to create such Email Things, regardless of which Gmail account these emails belong to.