Introducing rethink-migrate.
It will be run as part of npm postinstall, only in production.

There's also a way to run:
npm run upgrade-db
npm run downgrade-db

The name of the DB will be taken from the env variable RDB_DB_NAME.
It works for windows and linux envs.

Migrations are now created in the 'migrations' folder under root folder.