Introducing rethink-migrate again.
It will be run as part of npm postinstall, only in production.

There's also a way to run:
npm run upgrade-db (use '--' to provide arguments to the script)
npm run downgrade-db

Connection options will be taken from env variables.
It works for windows and linux environments.

Migrations are now created in the 'db/migrations' folder.
It works for remote DBs too, using a certificate and an auth key provided as env variable.