Introducing rethink-migrate again.
The only way to pass rethink-migrate the TLS certificate is in a JSON format, i.e. via a file.
rethink-migrate reads only the file 'database.json' in its root dir ('db' dir in our case).
But when this file is there with the CA, it's impossible to run it WITHOUT the CA, which is what we'd want to do in dev mode.
Originally I tried to put the database.json in a different folder ('config') and copy it to 'db' only in production.
But that didn't work, because 'bash cp' - which is banal way to copy files - didn't work.

So we ended up with this database.json file in the 'db' folder always, which would work for production, but would be annoying for dev.