Introducing All Things page.
Currently we show all Things, including from Email, Slack and other integrations.
There are no actions on the Things in the All page.
You can enter Thing Page from there, and you can still reply on the Thing.

We try to show information in the status to help you understand the context of the Thing.