freection-211 In Gmail, I can't Do an email from the full email page
Adding a Do button to the toolbar in Gmail.
It acts upon the current open thread.

Thing is, there's no way to disable it, since there's no convenient API in InboxSDK for changing the buttons in the toolbar.
Currently clicking Do takes you back to the inbox, but the button will still be enabled when re-entering the thread page.