freection-212 I don't have to enter a message when performing action on Thing Page
We go with the easiest solution currently - after clicking an action (Done, Dismiss, etc.), we open another tab in the message box, only for the sake of the action itself.
The text in this new message box is mandatory, so the user can't perform an action without writing a comment.