Bug fix - application not running due to babel presets problem.
I used .babelrc in order to support certain presets for mocha, which does not have an option to pass presets inline.
But the presets I put in .babelrc are now used throughout the system, in other babel commands not related to mocha.
This fucked up the build.
So now I run mocha indirectly, using babel-node, which accepts presets inline, and I remove the .babelrc.