Adding UT to Comment flow.
Refactored the thing-page-reducer along the way, so that it only handles actions of event updates. It does not handle thing command actions, and that's okay.
Problem with dealing with the thing command actions was that the corresponding endpoints in the server always return a Thing, and not an Event, and therefore it would be more difficult for the reducer to know what and how to update.