freection-260 Count of To Do and Follow Up is not updated after Do or sending new Thing
Pages would not get notifications, because they were in REQUIRE_UPDATE state.
This happened because once the application connects to the push mechanism, we immediately invalidate all states. We do that in order to make sure that the pages re-fetch everything on startup, since we might get new notifications in the time frame between fetching and connecting to the push mechanism.

Bug was that we invalidated the states, but didn't do a re-fetch immediately - we'd rather wait for the user to navigate a certain page in order re-fetch its state.
Now, we re-fetch all invalidated pages immediately - which is slower - but would put them in FETCHED state, which will enable them to get notifications.