Introducing a thank you page.
It will be served by our regular application, and will be redirected to from the site's landing page.
We write the thank you page here, and not using the same platform that we used for the landing page itself, because we need to inject a piece of JS code, for the sake of Facebook's pixel analytics. The instapage platform we used for the landing page requires us to pay 55$ monthly in order to edit JS code, so right now we'd rather just do it here.
We hope we can redirect and monitor the leads when the two pages are served in different domains.