Asana  - changing task to Done when completed event arrives.
I do that only for tasks that are in NEW, INPROGRESS or REOPENED status.
Otherwise, if the task is already "closed" from within Freection, I think there's no need to change it's status.
It could be buggy, for instance, if it's already in CLOSED status, and we then take it back to DONE status.