Asana  - adding unassign functionality.
There's a new event called UNASSIGNED which we create when an Asana task is unassigned.
The Freection user who is the unassignee will get a notification, with a Close action only, which will remove the task from his To Do page.

I refactored the ThingService.closeAck function, to be in line with how these functions are written, after we refactored them like one month ago.

Also, notice that I audit an event to Intercom, but only for the unassignee, which is guaranteed to be a Freection user.
The unassigner is an Asana user, and not necessarily a Freection user.
The unassigner will only be a field on the event of the unassignee.