freection-99 The scrolling on whats new is only on the preview items list which is annoying...
The position of the scroll bar itself is ugly in all of the pages with scroll (Inbox, To Do, Follow Up, All Items).
I make some changes to make it look a bit better.
They are less convenient architecture-wise, because the same style (of max-width) is duplicated among these pages.
But that's the only way I see to do it right now.