Asana - fetch user uncomplete tasks when enabling projects.
The tasks that are assigned to the user, and that are not complete, would appear in the user's To Do directly (not in the Inbox).
The user would get a desktop notification for the CREATED events, though.
The original creator of the task in Asana would be set as creator for the task in Freection.