Refactoring - adding userProfile state in addition to auth state.
Auth state is coupled to the user token, and is therefore something that's not easy to change.
We need the actual user info to be in a regular state, agnostic to the token.
Token, as a state, should be used not for user info, but for authentication status - isAuthenticated and such.